feat: add compile flags

+++ daemon/daemon.cpp
@@ -105,7 +105,7 @@ void Daemon::run(const std::filesystem::path & current_dir, const std::string & 
 
     while (true)
     {
-        for(int i=0; i < table.size(); ++i)
+        for(size_t i=0; i < table.size(); ++i)
         {
             auto now_time = std::chrono::steady_clock::now();
             auto prev_time = time_points[i];
